refactor: split up logic validation logic into separate util functions
