+++ packages/graphql/src/lib/validateSelectionOptions.ts
@@ -1,0 +1,47 @@
+import { QuestionOptionsArgs } from './validateAndProcessElementOptions.js'
+
+function validateSelectionOptions(options?: QuestionOptionsArgs | null) {
+  // options and hasSampleSolution need to be defined
+  if (
+    !options ||
+    typeof options.hasSampleSolution !== 'boolean' ||
+    options.hasSampleSolution === null
+  ) {
+    console.error(
+      'Options and sample solution flag are required for selection questions'
+    )
+    return false
+  }
+
+  // number of inputs needs to be specified and valid
+  if (
+    typeof options.numberOfInputs !== 'number' ||
+    options.numberOfInputs === null ||
+    options.numberOfInputs < 1
+  ) {
+    console.error('Number of inputs needs to be specified and valid')
+    return false
+  }
+
+  // answer collection needs to be defined for selection questions
+  if (
+    typeof options.answerCollection !== 'number' ||
+    options.answerCollection === null
+  ) {
+    console.error(
+      'Answer collection needs to be specified for selection questions'
+    )
+    return false
+  }
+
+  // if sample solution is activated, at the numberOfInputs sample solutions need to be defined
+  if (
+    options.hasSampleSolution &&
+    (!options.correctAnswers ||
+      options.correctAnswers.length < options.numberOfInputs)
+  ) {
+    return false
+  }
+}
+
+export default validateSelectionOptions
